backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.18
tensorflow==2.21.0
opencv-python>=4.10.0.84
numpy>=1.26.0
pillow>=10.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.30
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator>=2.1.0
